+++ src/main/java/com/practice/web/model/Role.java
@@ -15,7 +15,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name="id_role")
-    private int id_role;
+    private Long id_role;
 
     @Column(name="role_name")
     private String role;
