+++ src/main/java/com/practice/web/repository/BookCategoryRepository.java
@@ -3,6 +3,6 @@ package com.practice.web.repository;
 import com.practice.web.model.BookCategory;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookCategoryRepository extends CrudRepository<BookCategory, Integer> {
+public interface BookCategoryRepository extends CrudRepository<BookCategory, Long> {
 
 }
